Fix random check for bomb
The function repeats values due to certain values of position_x and position_y being duplicates, because adding them together can cancel them out, if one is negative. So instead, add the str representation of the digits to the str representation of the seed, then the new seed is the hash of that str. It must be hashed because MicroPython's random module, unlike the current release of CPython, only accepts ints.