Increase function performance
Before, every tile in checked tiles and flagged tiles would be iterated on. This
increased performance while both were less than the screen resolution. However,
this game is supposed to be optimized to go for as long as possible, so a
performance that decreases as the game goes on is undesired.